+++ spec/support/locale.rb
@@ -1,1 +1,1 @@
-I18n.available_locales = %w[en en-GB da de es fr it ja nl pl pt-BR sv tr uk vi zh-CN]
+I18n.available_locales = Dir[File.expand_path("../../config/locales/*.yml", __dir__)].map { |path| File.basename(path, ".yml") }